refactor(contexts): optimize performance and reorganize filter hooks
- Use state updater function pattern in RecentlyEditedContext for optimal performance
- Consolidate filter hooks imports from FilterProvider to FilterHooks
- Remove debug console statements from test files
- Enhance queryKeys documentation with detailed explanations and examples
- Eliminate unnecessary callback dependencies to prevent recreation
